refactor: move pdf report generation to new helper
