+++ lib/presentation/screens/home_screen.dart
@@ -1,12 +1,8 @@
 
 import 'dart:convert';
-import 'dart:io' show Directory, File, Platform;
-import 'dart:math';
-import 'dart:math' as math;
-import 'package:intl/intl.dart';
+import 'dart:io' show Platform;
 import 'package:dropdown_search/dropdown_search.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -14,9 +10,6 @@ import 'package:go_router/go_router.dart';
 import 'package:ndri_dairy_risk/presentation/screens/preview_answers_screen.dart';
 import 'package:ndri_dairy_risk/presentation/widgets/app_text.dart';
 import 'package:open_file/open_file.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:pdf/pdf.dart';
-import 'package:pdf/widgets.dart' as pw;
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../config/app_colors.dart';
 import '../../config/app_strings.dart';
@@ -29,6 +22,7 @@ import '../../logic/risk_assessment/bloc/risk_assessment_bloc.dart';
 import '../../logic/risk_assessment/bloc/risk_assessment_event.dart';
 import '../../logic/risk_assessment/bloc/risk_assessment_state.dart';
 import '../../logic/score_calculate/question_weight.dart';
+import '../utils/report_generator.dart';
 
 @pragma('vm:entry-point')
 void notificationTapBackground(
@@ -448,646 +442,6 @@ class _HomeScreenState extends State<HomeScreen> {
             ),
           ),
     );
-  }
-  pw.Widget gaugeScore({
-    required String label,
-    required String score,
-    required double value,
-    double size = 120,
-    PdfColor color = PdfColors.blue,
-  })
-  {
-    value = value.clamp(0.0, 1.0);
-    final level = _riskLevelFromValue(value);
-    return pw.Padding(
-      padding: const pw.EdgeInsets.symmetric(vertical: 2),
-      child: pw.Row(
-        crossAxisAlignment: pw.CrossAxisAlignment.center,
-        mainAxisAlignment: pw.MainAxisAlignment.spaceBetween,
-        children: [
-          pw.SizedBox(
-            width: 190,
-            child: pw.Text(
-              label,
-              style: pw.TextStyle(fontSize: 17, fontWeight: pw.FontWeight.bold),
-            ),
-          ),
-          pw.SizedBox(width: 6),
-          /*   pw.Container(
-              width: size,
-              height: size / 2,
-              child: pw.CustomPaint(
-                size: PdfPoint(size, size),
-                painter: (pw.Context ctx, pw.Canvas canvas, pw.Rect rect) {
-                  final radius = size / 2;
-                  final center = pw.Offset(radius, radius);
-                  final trackRadius = radius * 0.85;
-                  final segmentWidth = radius * 0.22;
-                  final colors = [
-                    PdfColors.blue,
-                    PdfColors.green,
-                    PdfColors.yellow,
-                    PdfColors.orange,
-                    PdfColors.red,
-                  ];
-                  final segmentAngle = pi / colors.length;
-                  for (var i = 0; i < colors.length; i++) {
-                    final paint = pw.Paint()
-                      ..style = pw.PaintingStyle.stroke
-                      ..strokeWidth = segmentWidth
-                      ..color = colors[i];
-                    final start = pi + i * segmentAngle;
-                    canvas.drawArc(
-                      pw.Rect.fromCircle(center: center, radius: trackRadius),
-                      start,
-                      segmentAngle,
-                      false,
-                      paint,
-                    );
-                  }
-                  final borderPaint = pw.Paint()
-                    ..style = pw.PaintingStyle.stroke
-                    ..strokeWidth = 1
-                    ..color = PdfColors.black;
-                  canvas.drawArc(
-                    pw.Rect.fromCircle(center: center, radius: trackRadius + segmentWidth / 2),
-                    pi,
-                    pi,
-                    false,
-                    borderPaint,
-                  );
-
-                  final pointerAngle = pi + value * pi;
-                  final pointerOffset = pw.Offset(
-                    center.dx + trackRadius * cos(pointerAngle),
-                    center.dy + trackRadius * sin(pointerAngle),
-                  );
-                  canvas.drawCircle(
-                    pointerOffset,
-                    segmentWidth / 4,
-                    pw.Paint()..color = PdfColors.white,
-                  );
-                  canvas.drawCircle(
-                    pointerOffset,
-                    segmentWidth / 4,
-                    pw.Paint()
-                      ..style = pw.PaintingStyle.stroke
-                      ..strokeWidth = 1
-                      ..color = PdfColors.black,
-                  );
-                },
-              ),
-            ),*/
-          pw.SizedBox(width: 10),
-          pw.Container(
-            width: 40,
-            alignment: pw.Alignment.centerLeft,
-            child: pw.Text(
-              score,
-              style: pw.TextStyle(fontSize: 16, fontWeight: pw.FontWeight.bold),
-            ),
-          ),
-          pw.SizedBox(width: 6),
-          pw.Container(
-            width: 50,
-            child: pw.Text(
-              level,
-              style: pw.TextStyle(
-                fontSize: 15,
-                color: _riskColor(level),
-                fontWeight: pw.FontWeight.bold,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-  pw.Widget _pointerCircle(double size) {
-    return pw.Container(
-      width: size,
-      height: size,
-      decoration: pw.BoxDecoration(
-        color: PdfColors.white,
-        shape: pw.BoxShape.circle,
-        border: pw.Border.all(color: PdfColors.blue, width: 2),
-      ),
-    );
-  }
-
-  pw.Widget imageScoreBarWithArrow({
-    required String label,
-    required String score,
-    required double value,
-    required pw.MemoryImage barImage,
-    double barWidth = 180,
-    double barHeight = 33,
-    String? level,
-    PdfColor? levelColor,
-  })
-  {
-    value = value.clamp(0.0, 1.0);
-    return pw.Padding(
-      padding: const pw.EdgeInsets.symmetric(vertical: 2),
-      child: pw.Row(
-        crossAxisAlignment: pw.CrossAxisAlignment.center,
-        mainAxisAlignment: pw.MainAxisAlignment.spaceBetween,
-        children: [
-          pw.SizedBox(
-            width: 190,
-            child: pw.Text(
-              label,
-              style:
-              pw.TextStyle(fontSize: 17, fontWeight: pw.FontWeight.bold),
-            ),
-          ),
-          pw.SizedBox(width: 6),
-          pw.Container(
-            width: barWidth,
-            height: barHeight + 20,
-            child: pw.Stack(
-              children: [
-                pw.Positioned(
-                  left: 0,
-                  top: 14,
-                  child:
-                  pw.Image(barImage, width: barWidth, height: barHeight),
-                ),
-                pw.Positioned(
-                  left: (barWidth - 22) * value,
-                  top: 0,
-                  child: _pointerCircle(22),
-                ),
-              ],
-            ),
-          ),
-          pw.SizedBox(width: 10),
-          pw.Container(
-            width: 40,
-            alignment: pw.Alignment.centerLeft,
-            child: pw.Text(
-              score,
-              style:
-              pw.TextStyle(fontSize: 16, fontWeight: pw.FontWeight.bold),
-            ),
-          ),
-          pw.SizedBox(width: 6),
-          pw.Container(
-            width: 50,
-            child: pw.Text(
-              level ?? '',
-              style: pw.TextStyle(
-                fontSize: 15,
-                color: levelColor ?? PdfColors.black,
-                fontWeight: pw.FontWeight.bold,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> _generateReport() async {
-    await _initPermissions();
-    final st = context.read<RiskAssessmentBloc>().state;
-    if (st is! RiskAssessmentLoaded) return;
-
-    final pdf = pw.Document();
-
-    final bgImage = pw.MemoryImage(
-      (await rootBundle.load('assets/images/ic_socio_climatic_dia.webp')).buffer.asUint8List(),
-    );
-    final barImage = pw.MemoryImage(
-      (await rootBundle.load('assets/images/hazard_bar.png')).buffer.asUint8List(),
-    );
-    final gaugeImage = pw.MemoryImage(
-      (await rootBundle.load('assets/images/socio_gauge.png')).buffer.asUint8List(),
-    );
-    final rainbowGaugeImage = pw.MemoryImage(
-      (await rootBundle.load('assets/images/rainbow_color.png')).buffer.asUint8List(),
-    );
-
-    String asFixed(dynamic val) {
-      if (val == null) return '0.00';
-      if (val is num) return val.toStringAsFixed(2);
-      if (val is String) {
-        return double.tryParse(val)?.toStringAsFixed(2) ?? '0.00';
-      }
-      return '0.00';
-    }
-
-    String hazardLevelFromValue(double v) {
-      if (v < 0.254952719) return 'Very Low';
-      if (v < 0.359426855) return 'Low';
-      if (v < 0.419781228) return 'Medium';
-      if (v < 0.512033541) return 'High';
-      return 'Very High';
-    }
-
-    String exposureLevelFromValue(double v) {
-      if (v < 0.3629) return 'Very Low';
-      if (v < 0.4252) return 'Low';
-      if (v < 0.4670) return 'Medium';
-      if (v < 0.5131) return 'High';
-      return 'Very High';
-    }
-
-    PdfColor riskColor(String level) {
-      switch (level.toLowerCase()) {
-        case 'very low':
-          return PdfColors.green;
-        case 'low':
-          return PdfColors.lightGreen;
-        case 'medium':
-          return PdfColors.yellow;
-        case 'high':
-          return PdfColors.orange;
-        case 'very high':
-          return PdfColors.red;
-        default:
-          return PdfColors.grey;
-      }
-    }
-
-
-    pw.Widget gaugeWithOverlay({
-      required double value, // 0..1
-      required String dateText,
-      required double score,
-      required pw.MemoryImage gaugeImage,
-      double width = 540,
-      double height = 220,
-      double pointerSize = 20,
-    }) {
-      // Adjust these to match your PNG dimensions and overlay requirements
-      final gaugeCenter = Offset(width / 2, height - 52);
-      final pointerPos = gaugeCenter; // place pointer at center of gauge
-
-      return pw.Container(
-        width: width,
-        height: height,
-        alignment: pw.Alignment.center,
-        child: pw.Stack(
-          alignment: pw.Alignment.center,
-          children: [
-            pw.Padding(padding: pw.EdgeInsets.only(left: 12),
-              child:
-              pw.Image(
-                gaugeImage,
-                width: width,
-                fit: pw.BoxFit.fitWidth,
-              ),
-            ),
-            pw.Positioned(
-              left: 0,
-              right: 0,
-              top: height * 0.33,
-              child: pw.Column(
-                mainAxisSize: pw.MainAxisSize.min,
-                crossAxisAlignment: pw.CrossAxisAlignment.center,
-                children: [
-                  pw.Text(
-                    dateText,
-                    textAlign: pw.TextAlign.center,
-                    style: pw.TextStyle(
-                      fontSize: 22,
-                      fontWeight: pw.FontWeight.bold,
-                    ),
-                  ),
-                  pw.SizedBox(height: 8),
-                  pw.Text(
-                    score.toStringAsFixed(2),
-                    textAlign: pw.TextAlign.center,
-                    style: pw.TextStyle(
-                      fontSize: 40,
-                      fontWeight: pw.FontWeight.bold,
-                      color: PdfColors.blue,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            pw.Positioned(
-              left: pointerPos.dx - pointerSize / 2,
-              top: 10,
-              child: _pointerCircle(pointerSize),
-            ),
-          ],
-        ),
-      );
-    }
-
-    String name = nameCtrl.text.trim();
-    String stateName = selectedState ?? '';
-    String district = selectedDistrict ?? '';
-    String block = blockCtrl.text.trim();
-    String village = villageCtrl.text.trim();
-
-    final answers = _localAnswers.map((k, v) => MapEntry(k, v.toString()));
-    // Compute and log vulnerability values similar to exposure
-    final vulnDetails = computeVulnerabilityDetails(answers);
-    final vulnVal = vulnDetails['score'] as double;
-    final vulnMap = vulnDetails['values'] as Map<String, double>;
-    final vulnValuesStr = vulnMap.entries
-        .map((e) => '${e.key}: ${e.value}')
-        .join(' + ');
-    print('Vulnerability values -> $vulnValuesStr');
-    print(
-        'Vulnerability details -> sum: ${vulnDetails['sum']!.toStringAsFixed(2)}, weight: ${vulnDetails['weight']!.toStringAsFixed(2)}, score: ${vulnDetails['score']!.toStringAsFixed(2)}');
-
-    final expDetails = computeExposureDetails(answers);
-    final expVal = expDetails['score'] as double;
-    final valMap = expDetails['values'] as Map<String, double>;
-    final valuesStr = valMap.entries
-        .map((e) => '${e.key}: ${e.value}')
-        .join(' + ');
-    print('Exposure values -> $valuesStr');
-    print(
-        'Exposure details -> sum: ${expDetails['sum']!.toStringAsFixed(2)}, weight: ${expDetails['weight']!.toStringAsFixed(2)}, score: ${expDetails['score']!.toStringAsFixed(2)}');
-
-    String vulnerabilityScore = vulnVal.toStringAsFixed(2);
-    String exposureScore = expVal.toStringAsFixed(2);
-    String getTotalScore = asFixed(vulnerabilityScore).toString() + asFixed(exposureScore);
-    String getHazardScore = asFixed(hazardLevelFromValue(200.10));
-    print('getted the value $getTotalScore');
-
-    double hazardVal = LocationService().hazardFor(district);
-    String hazardScore = hazardVal.toStringAsFixed(2);
-    String riskScore = asFixed(st.answers['riskScore']);
-    String riskLevel = st.answers['riskLevel'] ?? 'Moderate';
-    print('getted the hrisk scrore $riskLevel');
-
-    pw.Widget buildLegendRow(PdfColor color, String label) {
-      return pw.Row(
-        mainAxisAlignment: pw.MainAxisAlignment.start,
-        children: [
-          pw.Container(
-            width: 14,
-            height: 14,
-            decoration: pw.BoxDecoration(
-              color: color,
-              border: pw.Border.all(color: PdfColors.black, width: 0.5),
-            ),
-          ),
-          pw.SizedBox(width: 8),
-          pw.Text(
-            label,
-            style: pw.TextStyle(
-              fontSize: 14,
-              color: PdfColors.black,
-            ),
-          ),
-        ],
-      );
-    }
-
-    pdf.addPage(
-      pw.Page(
-        pageFormat: PdfPageFormat.a4,
-        margin: pw.EdgeInsets.zero,
-        build: (pw.Context context) {
-          final hazardLevel = hazardLevelFromValue(hazardVal);
-          final hazardValueForBar = hazardVal.clamp(0.0, 1.0);
-          final exposureLevel = exposureLevelFromValue(expVal);
-          final exposureValueForBar = expVal.clamp(0.0, 1.0);
-
-          return pw.Stack(
-            children: [
-              pw.Positioned.fill(
-                child: pw.Opacity(
-                  opacity: 0.10,
-                  child: pw.Image(bgImage, fit: pw.BoxFit.contain),
-                ),
-              ),
-              pw.Padding(
-                  padding: const pw.EdgeInsets.symmetric(horizontal: 20, vertical: 5),
-                  child: pw.Column(
-                    children: [
-                      pw.Container(
-                        width: double.infinity,
-                        height: 90,
-                        decoration: pw.BoxDecoration(
-                          color: PdfColor.fromHex('#009688'),
-                          borderRadius: const pw.BorderRadius.only(
-                            bottomLeft: pw.Radius.circular(100),
-                            bottomRight: pw.Radius.circular(100),
-                          ),
-                        ),
-                        alignment: pw.Alignment.center,
-                        child: pw.Column(
-                          mainAxisAlignment: pw.MainAxisAlignment.center,
-                          children: [
-                            pw.Text('Socio-climatic Risk of',
-                                style: pw.TextStyle(
-                                    fontSize: 32,
-                                    fontWeight: pw.FontWeight.bold,
-                                    color: PdfColors.white)),
-                            pw.Text('Smallholder Dairy Farmer',
-                                style: pw.TextStyle(
-                                    fontSize: 24,
-                                    fontWeight: pw.FontWeight.bold,
-                                    color: PdfColors.yellow)),
-                          ],
-                        ),
-                      ),
-                      pw.SizedBox(height: 10),
-                      pw.Padding(
-                        padding: const pw.EdgeInsets.symmetric(horizontal: 20),
-                        child: pw.Column(
-                          crossAxisAlignment: pw.CrossAxisAlignment.start,
-                          children: [
-                            pw.Row(children: [
-                              pw.Text('Name of the dairy farmer: ',
-                                  style: pw.TextStyle(
-                                      fontSize: 20,
-                                      fontWeight: pw.FontWeight.normal)),
-                              pw.Text(name,
-                                  style: const pw.TextStyle(
-                                      fontSize: 18,
-                                      color: PdfColors.blueAccent))
-                            ]),
-                            pw.SizedBox(height: 10),
-                            pw.Row(children: [
-                              pw.Text('State: ',
-                                  style: pw.TextStyle(
-                                      fontSize: 20,
-                                      fontWeight: pw.FontWeight.bold,
-                                      color: PdfColors.brown400)),
-                              pw.Text(stateName,
-                                  style: const pw.TextStyle(
-                                      fontSize: 18,
-                                      color: PdfColors.blueAccent)),
-                              pw.Spacer(),
-                              pw.Text('Block: ',
-                                  style: pw.TextStyle(
-                                      fontSize: 20,
-                                      fontWeight: pw.FontWeight.bold,
-                                      color: PdfColors.brown400)),
-                              pw.Text(block,
-                                  style: const pw.TextStyle(
-                                      fontSize: 18,
-                                      color: PdfColors.blueAccent)),
-                            ]),
-                            pw.SizedBox(height: 5),
-                            pw.Row(children: [
-                              pw.Text('District: ',
-                                  style: pw.TextStyle(
-                                      fontSize: 20,
-                                      fontWeight: pw.FontWeight.bold,
-                                      color: PdfColors.brown400)),
-                              pw.Text(district,
-                                  style: const pw.TextStyle(
-                                      fontSize: 18,
-                                      color: PdfColors.blueAccent)),
-                              pw.Spacer(),
-                              pw.Text('Village: ',
-                                  style: pw.TextStyle(
-                                      fontSize: 20,
-                                      fontWeight: pw.FontWeight.bold,
-                                      color: PdfColors.brown400)),
-                              pw.Text(village,
-                                  style: const pw.TextStyle(
-                                      fontSize: 18,
-                                      color: PdfColors.blueAccent)),
-                            ]),
-                          ],
-                        ),
-                      ),
-                      pw.SizedBox(height: 10),
-                      imageScoreBarWithArrow(
-                        label: '1. Vulnerability score',
-                        score: vulnerabilityScore,
-                        value: double.tryParse(vulnerabilityScore) ?? 0.0,
-                        barImage: barImage,
-                        level: hazardLevelFromValue(
-                            double.tryParse(vulnerabilityScore) ?? 0.0),
-                        levelColor: riskColor(hazardLevelFromValue(
-                            double.tryParse(vulnerabilityScore) ?? 0.0)),
-                      ),
-                      pw.SizedBox(height: 10),
-                      imageScoreBarWithArrow(
-                        label: '2. Exposure score',
-                        score: exposureScore,
-                        value: exposureValueForBar,
-                        barImage: barImage,
-                        level: exposureLevel,
-                        levelColor: riskColor(exposureLevel),
-                      ),
-                      pw.SizedBox(height: 10),
-                      imageScoreBarWithArrow(
-                        label: '3. Hazard score',
-                        score: hazardScore,
-                        value: hazardValueForBar,
-                        barImage: barImage,
-                        level: hazardLevel,
-                        levelColor: riskColor(hazardLevel),
-                      ),
-                      pw.SizedBox(height: 10),
-                      pw.Center(
-                        child: gaugeWithOverlay(
-                          pointerSize: 20,
-                          value: double.tryParse(riskScore) ?? 50.0,
-                          dateText: DateFormat('dd MMM yyyy').format(DateTime.now()),
-                          score: double.tryParse(riskScore) ?? 50.0,
-                          gaugeImage: rainbowGaugeImage,
-                          width: 500,
-                          height: 220,
-                        ),
-                      ),
-                      pw.SizedBox(height: 10),
-                      pw.RichText(
-                        text: const pw.TextSpan(
-                          style: pw.TextStyle(fontSize: 16),
-                          children: [
-                            pw.TextSpan(
-                              text: 'Disclaimer: ',
-                              style: pw.TextStyle(
-                                color: PdfColors.blue,
-                              ),
-                            ),
-                            pw.TextSpan(
-                              text:
-                              'Above socio-climatic risk score is calculated based on information provided by the farmer.',
-                              style: pw.TextStyle(
-                                color: PdfColors.black,
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ],
-                  )),
-            ],
-          );
-        },
-      ),
-    );
-
-
-    Directory downloadsDir;
-    if (Platform.isAndroid) {
-      downloadsDir = Directory('/storage/emulated/0/Download');
-    } else {
-      downloadsDir = (await getDownloadsDirectory())!;
-    }
-    if (!await downloadsDir.exists()) {
-      await downloadsDir.create(recursive: true);
-    }
-
-    final fileName = 'report_${DateTime.now().millisecondsSinceEpoch}.pdf';
-    final outFile = File('${downloadsDir.path}/$fileName');
-    await outFile.writeAsBytes(await pdf.save());
-
-    const androidDetails = AndroidNotificationDetails(
-      'reports',
-      'Reports',
-      channelDescription: 'Your report is ready to open',
-      importance: Importance.high,
-      priority: Priority.high,
-    );
-    await _localNotifications.show(
-      0,
-      'Report saved',
-      'Tap to open your PDF',
-      const NotificationDetails(android: androidDetails),
-      payload: outFile.path,
-    );
-  }
-
-
-  String _riskLevelFromValue(double v) {
-    if (v < 0.2) return 'Very Low';
-    if (v < 0.4) return 'Low';
-    if (v < 0.6) return 'Moderate';
-    if (v < 0.8) return 'High';
-    return 'Very High';
-  }
-
-  PdfColor _riskColor(String level) {
-    switch (level.toLowerCase()) {
-      case 'very low':
-        return PdfColors.green;
-      case 'low':
-        return PdfColors.lightGreen;
-      case 'moderate':
-        return PdfColors.orange;
-      case 'high':
-        return PdfColors.red;
-      case 'very high':
-        return PdfColors.red900;
-      default:
-        return PdfColors.grey;
-    }
-  }
-
-  String riskLevelFromScore(String score) {
-    double val = double.tryParse(score) ?? 0.0;
-    if (val >= 0.8) return 'Very High';
-    if (val >= 0.6) return 'High';
-    if (val >= 0.4) return 'Moderate';
-    if (val >= 0.2) return 'Low';
-    return 'Very Low';
   }
 
   Future<void> _loadLocalAnswers() async {
@@ -3189,7 +2543,17 @@ class _HomeScreenState extends State<HomeScreen> {
                         ),
                       ),
                   child: ElevatedButton.icon(
-                    onPressed: _generateReport,
+                    onPressed: () => ReportGenerator.generate(
+                      context: context,
+                      answers: _localAnswers,
+                      notifications: _localNotifications,
+                      initPermissions: _initPermissions,
+                      name: nameCtrl.text.trim(),
+                      block: blockCtrl.text.trim(),
+                      village: villageCtrl.text.trim(),
+                      stateName: selectedState,
+                      district: selectedDistrict,
+                    ),
                     icon: TweenAnimationBuilder<double>(
                       tween: Tween(begin: 0.0, end: 1.0),
                       duration: const Duration(milliseconds: 1100),
